removed code duplication from parsers
- new abstract base class for XyzOntologyTransformers
- changed JavaOntologyTransformer to extend the new class
- changed KotlinOntologyTransformer to extend the new class
