+++ org.architecture.cnl.parent/org.architecture.cnl/.classpath
@@ -14,6 +14,13 @@
 			<attribute name="maven.pomderived" value="true"/>
 		</attributes>
 	</classpathentry>
+	<classpathentry kind="src" output="target/test-classes" path="src/test/xtend-gen">
+		<attributes>
+			<attribute name="test" value="true"/>
+			<attribute name="optional" value="true"/>
+			<attribute name="maven.pomderived" value="true"/>
+		</attributes>
+	</classpathentry>
 	<classpathentry kind="src" output="target/test-classes" path="src/test/xtext-gen">
 		<attributes>
 			<attribute name="test" value="true"/>
